Very basic demo working
The skeleton of the demo is working. It currently has customizable thrust curves, fourth-order Runge Kutta implementation for accuracy of the path, and stats/plots of the path of the rocket and the thrust curves. This can be run through user-interaction.py provided the other python files are in the same folder. I'll focus on a Kalman filter for simulated noise of the rocket as it ascends next.