Kalman Filter Addition
Added a basic 1D linear kalman filter to the rocket's path, as well as simulated noise for it to process. The denoised path doesn't seem very smooth, so I will attempt to improve the filter or smooth it out with another algorithm. The kalman filter functions can be found within the kalmanfilter.py file, and the KalmanFilterPlay.ipynb isn't part of the program and is just some examples I was running through to get an idea of how the filter works.