RTS Smoothing + Docstring updates
In this update I have implemented RTS Smoothing on the Kalman filter in 1 dimension (Vertical/Z) for the rocket's path. This takes noise, filters it, and smooths the path out so it moreso reflects the actual path. This is done to simulate real environments with noise. I have also updated some of the docstrings for some of the functions in this program.